requirements.txt
langchain==0.3.27
langchain-community==0.3.27
GitPython
chromadb
openai
python-dotenv
streamlit
tiktoken